[Game Systems : Serialization] Make World Serializable
The system is as follows:  Everything in the world has a SerializableX sister class.  That sister class inherits from ISerializable<X>, has a constructor of X, contains all of the variables necessary to rebuild X in the exact state it came out (or close enough), and can turn itself back into an X.

This system has a few advantages.

1) The JSON is abstracted a bit from the object itself, so in-memory representations can change easily, and we don't ever have to worry about breaking the JSON without a compiler warning saying so.

2) We have an intermediary layer that can handle any future breaking changes in the data, post-release.

3) The outputs of the sister class mirror the outputs of the builder path for the same object.  I.E., they both create the object from the same constructor.  This allows the constructor to remain deterministic (putting any random generation on the builder), and allows us to strip out any unnecessary data or processing and know exactly which variables to maintain in the JSON

4) No JSON in random classes!  Almost.  IntVector2 has a gross one.

5) It's pretty fast.

Also, the GameSave class acts as the object to store all game save data, and can be told about objects as they become dirty, so we never save something that doesn't need to be saved.  Each chunk and its relevant data is output into its own file, and can be read in-or-out as necessary.  Right now, we just read everything in at once, but...