[Actors] Make Events for Death and Damage
We want to separate the logic for presentation and responses to death out from the ActorData.  It only know when to Die, not what that looks like, and it can vary wildly on different actors.  Much like the effects cueing system, we kinda just want whatever components that are on the thing to handle it dying.