[UI : Effects] Implement UIEffects System
UI effects are intended to be wrapped actions that can either be played instantly, or can be manipulated by other effects to make timing and error free editing of UI more reasonable.  The big problems with UI effects, normally, are:

1) They take up amounts of time which other effects or outcomes may not be aware of
2) They require complicated callback systems to react to or chain off of
3) They all have different values and types, making a Tween-like system complicated to implement

Instead, we will just say that any UI that has the ability to do an effect (either through a component, or through its own native code) will not execute the effect when asked, but will instead encapsulate it with a Play and OnComplete function.  Timing is simple, as we can just wait for the previous effect to raise OnComplete, and then play the next one.

TODO: Separate as many effects from the component logic as is reasonable, to make them more component based and interchangeable.