[Graphics] Generate Sprites For Blocks From Single Large Texture
We want the world to feel continuous and not have tiled or misaligned blocks.  So what we do is, make a huge texture (the size of a chunk), and then dynamically create the sprites from a section of it at runtime, based on their world position.  This creates some performance problems, but those can be solved by caching a sprite at position X, Y for a given texture (since it will potentially be reused every chunk), pre-loading the textures, and by finally adding that coroutine loading I've intended for a while where we don't load entire chunks in a single frame.  I guess that is the next thing to tackle...