[WorldGeneration : ChunkActivation] Create Distinct States of Chunk Activation:
Chunks are needed ASAP when requested, but it takes time load / build all of the constituent parts.  So let's go ahead and create the chunk immediately but mark it as not ready to use (with logic for handling interactings while not ready).  Then, let the hard work happen in a coroutine to avoid blocking, and activate the chunk when ready (with logic to process those interactions we couldn't handle).