[World Generation] Differentiate "Props" and "Features"
Props are generic, largely static objects that can be placed in the world easily via PropTypes. Ex: Wall Light, Rubble, Background Sprites

Features are special compenents of a space or of spaces that must be especially constructed to fit the situation, and generally provide some differentiation for that space type.  E.G. ElevatorShaft vs Shaft